"Update CustomToolbar and ToysLanding components to toggle between grid and map view using props and state"
This diff shows changes in CustomToolbar.jsx and ToysLanding.jsx files. In CustomToolbar.jsx, the code is refactored to accept viewType as a prop and update the state using the setViewType function. The ToggleButtonGroup's value is also updated to use the viewType prop. In ToysLanding.jsx, the CustomToolbar component is updated to pass the viewType state and setViewType function as props. The commented out Grid component in ToysLanding.jsx suggests that the code to render toys in grid view has been removed, but it is still present in the previous commit.