"Updated Navbar height, adjusted layout of ToyListMap and ToysLanding components"
This commit includes changes to the Navbar component by setting the height to "86px". Additionally, the layout of the ToyListMap and ToysLanding components has been adjusted by updating the Grid and Box components. The Box component in ToysLanding now has a width of "100%" and a height of "100vh". The Grid container in ToysLanding has also been updated to have a spacing of 0 and the margin has been removed. The ToyListMap component is now wrapped in a Grid item and the ToyCard components in ToysLanding are now wrapped in a Grid item with spacing 2.