"Update CustomToolbar to only switch view type if currently not set or switching from map to module view"
This commit updates the `CustomToolbar` component to only switch the `viewType` if it is currently not set or if the user is switching from the "map" view to the "module" view. This ensures that the view type is not unnecessarily toggled when the user is already in the desired view.