"Update ToyCard component to display zip code instead of location; fetch toys data from API instead of using static data"
This commit updates the `ToyCard` component to display the `zip_code` prop instead of the `location` prop. It also removes the use of static toys data in `ToysLanding` and replaces it with fetching toys data from an API. The `ToyCard` component has also been updated to use a default image instead of the toy's image. The `ToyList` component has been updated to pass the `zip_code` prop instead of the `image` and `location` props to the `ToyCard` component.